Allow Snow/Lucas VE1 and Spring to have different wakeup_input_device_names
Lucas VE1 has the right FW/config to suppress unwanted wakeup from TP.
For Spring, we have the right FW/config in M33, but not M31/M32. So
we can enable TP wakeup for M33, but not M31/M32 for now.

Lets have a different wakeup_input_device_names for Snow/Lucas VE1
and Spring so we can configure them differently.

This is a cherry-pick, which enables TP wakeup for Snow/Lucas VE1,
but not Spring.

BUG=chrome-os-partner:20940
TEST=emerge-daisy chromeos-bsp-daisy and
     emerge-daisy_spring chromeos-bsp-daisy
     make sure wakeup_input_device_names file exists

Change-Id: Ib7ff8a4459b091e2a14f163ef1a7f5969f975b44
Reviewed-on: https://chromium-review.googlesource.com/177091
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Queue: Yufeng Shen <miletus@chromium.org>
Tested-by: Yufeng Shen <miletus@chromium.org>