Disable FW Palm supression for Lucas VE1
The Palm supression block causes problems for real finger contacts on
the pad when a palm is detected.  It locks the whole pad up sometimes
and won't allow any movement until the palm is lifted.  We want the pad
to still work in there is a palm on part of the pad.

In another CL, I've re-tuned the palm detection thresholds in CMT to
work better with the 336s pad and it does a good job rejecting palms
without the FW's help now.  This is more flexible than having the FW do
it and doesn't suppress regular fingers when a palm is detected

BUG=chrome-os-partner:24045
TEST=This config was loaded onto the Lucas VE1 TP and it updated the config no
problem.  Usage was good when tested with this config and the CMT palm
detection setting updated

Original-Change-Id: Ia13045441fdba071db4b277049d6ff03e0ab8287
Originally-Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Originally-Reviewed-on: https://chromium-review.googlesource.com/178109
Originally-Reviewed-by: Yufeng Shen <miletus@chromium.org>
(cherry picked from commit 9a500c2ae036ce1926b0bd2e1d73b37233d307d1)

Change-Id: Ia9ad9634bc98534f1681bc8d978fbc68031b3a33
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/179286
Reviewed-by: Yufeng Shen <miletus@chromium.org>