Update Daisy VE1 touchpad config with extra noise immunity
This config was designed by Atmel Korea to address an electrical noise
issue that was being reported in Brazil.  As a side effect it seemed to
significantly mitigate another existing issue with 336s's touchpad.
There was an issue on 336s where the cursor would sometimes jump when
you
had one finger already on the pad and introduced a second.  The two
fingers would merge briefly before splitting, which could result in a
jumping cursor.  This config mitigates the problem, making it harder to
trigger.
BUG=chrome-os-partner:26299
TEST=manually tested the config on 336s

Change-Id: I869fcd604ea3e9e3e7c7b96ae9b6488d65dd16e8
Reviewed-on: https://chromium-review.googlesource.com/194553
Reviewed-by: Jongpil Jung <jongpil19.jung@samsung.com>
Tested-by: Jongpil Jung <jongpil19.jung@samsung.com>
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Commit-Queue: Charles Mooney <charliemooney@chromium.org>