mobbuild: Specify bootstrap directory to use for local cbuildbot.
By default cbuildbot bootstraps in the /tmp directory. On mobbuild /tmp
is mounted with noexec, so cbuildbot can't bootstrap out of it.
This CL creates a bootstrap directory in the local_buildtools work area,
and instructs cbuildbot to use that directory through the wrapper script
provided.

BUG=chromium:490935
TEST=cbiuldbot --test-bootstrap on a mobbuild device.

Change-Id: I9989053506fe186586f468eb5dd02ee90379cf16
Reviewed-on: https://chromium-review.googlesource.com/275143
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>