mobbuild: Add support for local hosts mappings.
mobbuild needs custom hosts mapping(s) so that the buildbot slave
running locally can communicate with the buildbot master.

We support this like we support all other site specific / credential
data, via the credentials disk.

To actually add the host mapping, we run dnsmasq listening on the
loopback device, and instruct shill to use that as an extra DNS server.

BUG=chromium:450333
TEST=`ping master2.golo.chromium.org`
     `ping chromium.googlesource.com`
TEST=reboot lakitu_mobbuild, ensure that shill/dnsmasq jobs are stable
     and work as expected w/ and w/o updates the nameservers.

Change-Id: I7d07e58e62a2d2f4b46ace26eece9967fe026e10
Reviewed-on: https://chromium-review.googlesource.com/292442
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>