Add cortex-a53 erratum fix board oak and arm64-generic (again).
Oak and arm64-generic boards are 64-bit kernel w/ 32-bit user
space. That means we build kernel using aarch64 toolchain, and other
packages using armv7a toolchain. We add this erratum fix option to all
packages. In theory, such option is not valid for 32-bit arm compiler
and linker, however, both are lenient enough to ignore silently the
erratum-fix option.

For Chrome in ChromeOS, it is a 32-bit application, and the errata
options break ChromeSDK builder, so I had another CL:299791 that filters
the options.

(This is a re-open of a previous reverted CL:294742. We had CL:299791 to
address the error caused by CL:294742)

BUG=chrome-os-partner:39170
CQ-DEPEND=CL:299791
TEST=cbuildbot oak-full

Change-Id: I7963c41983036e30d94fe69739ee82bcde969e11
Reviewed-on: https://chromium-review.googlesource.com/299704
Commit-Ready: Han Shen <shenhan@chromium.org>
Tested-by: Han Shen <shenhan@chromium.org>
Reviewed-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>