Sumo: only report KEY_POWER from internal "keyboard"
This system does not have internal keyboard, except that it exposes
power button as regular key on a PS/2 keyboard (in addition to the
standard ACPI power button). This is done so that we can implement
"fade" effect, since ACPI button does not send "release" event so we
can't see if and when the button was released when using ACPI event
device.

There is no way to query characteristics of keyboard attached to a
PS/2 port, so atkbd driver assumes that it always deals with full
standard keyboard, which causes upper layers to think that we indeed
have full keyboard always attached to the device and it never brings up
on-screen keyboard.

To work around this issue we implement custom keymap that we load
into our "keyboard" that disables all keys but KEY_POWER one.

BUG=chrome-os-partner:43385
TEST=Boot Sumo system without USB keyboard attached, verify that virtual
keyboard apperars on screen when user input is required.

Change-Id: I21fa96cf3a2d0896cb5a1b070aa2d008c1d5c3fa
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/305281
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Michael Spang <spang@chromium.org>