Ninja: do not advertise power button as full keyboard
This system does not have internal keyboard, except that it exposes
power button as regular key on a PS/2 keyboard (in addition to the
standard ACPI power button). This is done so that we can implement
"fade" effect, since ACPI button does not send "release" event so we
can't see if and when the button was released when using ACPI event
device.

There is no way to query characteristics of keyboard attached to a
PS/2 port, so atkbd driver assumes that it always deals with full
standard keyboard, which causes upper layers to think that we indeed
have full keyboard always attached to the device and so it does not
prompt user to connect USB keyboard to the unit.

To work around this issue we implement custom keymap that we load
into our "keyboard" that disables all keys but KEY_POWER one and also
install udev rule that resets ID_INPUT_KEYBOARD property that udev
assigns to the input device by default.

BUG=chrome-os-partner:42143
TEST=Boot Ninja system without USB keyboard attached, verify that virtual
keyboard appears on screen when user input is required.

Change-Id: I73e222b2d928521f04ce3b59f8dc02bccf84724f
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/305467
Reviewed-by: Michael Spang <spang@chromium.org>
(cherry picked from commit 5de596666ef7d2212f5832772406093859cec660)
Reviewed-on: https://chromium-review.googlesource.com/306677
Reviewed-by: Benson Leung <bleung@chromium.org>