Buddy: Use the right side DMIC pair for 2ch capture
Connect DMIC2 to stereo1 and DMIC1 to mono, so that
"DMIC Pair2" (on the right side of the camera) is
used when doing only 2ch capture.

The codec IF1 ADC1/2 Swap Mux are changed so that
left/right channel are ordered correctly. The recorded
file is verifed with Audacity. Broadwell TDM seems
to swap left and right channel in each DMIC pair.

MicPositions of "DMIC Pair2" is added.

BUG=chrome-os-partner:46725
TEST=2ch and 4ch recording on Buddy, see DMIC Pair2 is in slot0/1
and DMIC Pair1 is in slot2/3. Headset mic continues to work.

Change-Id: Iad23c105ac57050ffb966d758fbc3f02feab8181
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/310555
Reviewed-by: Alejandro Luebs <aluebs@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>