veyron: turning on screen rendering by depthcharge
This change turns on bitmap_in_cbfs, which will force vboot to let depthcharge
render screens using cbgfx. bmpblk.bin will not be included in GBB. Thus, GBB
size can be reduced.

The new broken screen (chromium:501060) is currently rendered in text only for
debugging. As a result of the change the proper screen will be effective.

BUG=chromium:502066
BRANCH=tot
TEST=Tested on Jerry

Change-Id: Ia6a9d1ea4596d64479108d03e578d5182205c1cc
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/313813
Reviewed-by: Julius Werner <jwerner@chromium.org>