Sumo: adjust to new udev syntax, enable brightness control keys
The new udev version changed syntax for dmi-based keyboard quirks, let's
adjust to the new syntax.

Also, Sumo has 2 buttons that supposed to control brightness of the
screen, let's map them to KEY_F6 and KEY_F7 that are used to control
brightness on Chrome OS devices.

BUG=b:35557186,b:35566210
TEST=Boot Sumo, verify brightness buttons work, verify virtual keyboard
is shown when external keyboard is disconnected.

Change-Id: Iaa4fb7c25d5dc8f405a4fbf05bb986f33fc800be
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/321121
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>