Update Chell's overlay to include touchpad hover
This CL adds a file to the powerd_prefs called detect_hover to
configure powerd to support a touchpad that can report hover.
This is used because Chell has a keyboard backlight, and if a
laptop has a hover-enabled touchpad and a keyboard backlight,
then powerd should be able to use the touchpad's "hover state"
to control the keyboard backlight before you even touch the
device.

BUG=chrome-os-partner:48920
TEST=built and deployed on Chell, and the file was correctly
install, activating the touchpad hover functionality in powerd

Change-Id: Ie4c9d832255975a4a044f741142102b93abc40af
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/323740
Reviewed-by: Dan Erat <derat@chromium.org>