overlay-beaglebone,-raspberrypi,-sama5d3: Specify partitioning label
util-linux 2.27.1 has support for both gpt and mbr partitioning.
When sfdisk is used to restore the mbr partitioning, the device is
probed for the partitioning type and gpt is found. This confuses
the restore process which expects guid gpt partition types but
MBR types are provided.

BUG=chromium:583671
TEST=Built image for beaglebone
CQ-DEPEND=CL:325680

Change-Id: I36ff7228f398adfe7944d05540106b85a3cb490d
Reviewed-on: https://chromium-review.googlesource.com/333903
Commit-Ready: Sabin Floares <sabin.floares@intel.com>
Tested-by: Sabin Floares <sabin.floares@intel.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>