Sentry: UCM config: Set Mic Volume to 24dB
The default value of "Mic Volume" control is 0dB which
is too low. This patch sets the gain to +24dB.

This patch also binds the "Mic Volume" control to the
sklnau8825max Headset Jack, so that CRAS can adjust it
at runtime for AGC.

BUG=chrome-os-partner:51527
TEST=The remote side can hear clearly in hangouts.
See "cras_test_client --capture_gain ..." affects "Mic Volume".

Change-Id: I5357ceb3ed04fdf19f7fe54bcdd4eeeb571cab70
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/334461
Reviewed-by: Dylan Reid <dgreid@chromium.org>