freon: actually disable tty2 for gru/kevin
For some overlays, ${TTY_CONSOLE} is initialized to tty2, and even
though we allegedly mask it here with adding "-tty_console_tty2" to USE.

This seems to be partially related to the old hack which is in use in
the freon overlay, where we only mask the USE flag and not the actually
TTY_CONSOLE variable. Now that portage has been patched to obviate the
hack, let's do this right: disable tty2 in the TTY_CONSOLE variable in
freon, and have subsequently-stacked overlays configure additional
console(s) in their make.defaults.

This only converts gru for now, as that's where I'm having trouble. We
could also do this conversion for other overlays now too.

BUG=chromium:606100
TEST=`emerge chromeos-base/tty` and check for /etc/init/console-*.conf

Change-Id: I66560f53549684e54dc82003b183ed4917d73841
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/340286
Reviewed-by: Mike Frysinger <vapier@chromium.org>