Peach: Stop building bootimage from ToT
Peach pit runs on the RO firmware, which is old and very different from the
current RO code. Actual firmware images used for the tests are separately built
and uploaded to a server. Therefore, it's better for CQ not to emerge it from
ToT.  This will avoid slowing down development unnecessarily.

BUG=chrome-os-partner:54813
BRANCH=tot
TEST=emerge-peach_pit virtual/target-chromium-os doesn't emerge
chromeos-bootimage

Change-Id: Icf39a6ca55d63abfb08394f16329aca0913014f4
Reviewed-on: https://chromium-review.googlesource.com/358670
Commit-Ready: Daisuke Nojiri <dnojiri@chromium.org>
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>