HACK: kevin: add vboot_reference USE flag for tpm2
Temporary hack to enable recovery images while tpm2 support
is not globally turned on for the image. Without it tpmc will
not work. To be moved to proper tpm2-specific overlays once
they are created and eventually reverted.

BUG=chrome-os-partner:55250
TEST=emerge-kevin vboot_reference, cros deploy to target,
verify that tpm2 version was built and tpmc works

Change-Id: I37770426b0891ece52254290661175b111dedeb2
Reviewed-on: https://chromium-review.googlesource.com/362393
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Philip Chen <philipchen@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>