baseboard-reef: Fix root path
root-path was incorrect in this file, but the error was masked by
CL:*265846.
That CL has been reverted by CL:*272139, fix root path where it should.

BUG=chrome-os-partner:54498
TEST=Look at DEFAULT_ROOTDEV in /images/reef/latest/partition_script.sh
With CL:*265846: check it is ..0000:00:1c.0/mmc_host/mmc1.. as expected.
With CL:*272139: notice it has changed to ...80860F14:00/mmc_host/mmc0..
With this change: verify it is back to the original value.

Change-Id: I0cb9cf309b07b844918e94e1dedf1091b5c246c4
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/365130
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>