kevin: Use Headphone control to control volume
Fix the MixerName in UCM Headphone section so we can use Headphone
control to change volume or mute.
With this we can control volume using control exposed by codec rather
than using software volume.
This switch does not cause noise when user mutes headphone during
playback.

Keep Headphones setting in enable/disable sequence so we can save power on
headphone path when it is not used.
The enable and disable sequences are updated to latest suggestion from
vendor.

BUG=chrome-os-partner:57586
TEST=change volume and see "Headphone" control has different volume
TEST=mute/unmute during headphone playback, there is no noise.
TEST=select/unselect during headphone playback, there is no noise.

Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Change-Id: I821f411ed09b06f9ab935a59c5e84e09e6402bc9
Reviewed-on: https://chromium-review.googlesource.com/389073
Tested-by: Chinyue Chen <chinyue@chromium.org>