power: add low_battery_shutdown_percent pref
The EC will do a hard shutdown at 3% to preserve battery power for
itself, and the powerd preference to shut down with 180 seconds
left on the battery doesn't reach that threshold, so powerd won't
shut down the computer before the EC decides to. This has implications
for tests that expect to run out the battery like power_LoadTest.

BUG=chrome-os-partner:58905
TEST=build image, check rootfs structure, check powerd logs for
"...and using low battery percent threshold of 4"

Change-Id: I37febf9bfe8af7b205bcd0ffc9c1a64954d0e277
Reviewed-on: https://chromium-review.googlesource.com/403474
Commit-Ready: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Tested-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>