Waiting for USB ethernet dongle without stopping the shill.
The upstart script moblab-network-bridge-init.conf will block 5 minutes
to wait for the USB ethnet dongle ready. We used to stop the shill
during the wait. This cause the Moblab to have no external interet
connection during the time. Also if the dongle is not configured, shill
is not restarted. We fix this by only stop the shill when USB dongle is
detected.

BUG=chromium:630034
TEST=manual.

Change-Id: I72a3ffd99a3dadf06c6bb042a5a19b02645ecd5d
Reviewed-on: https://chromium-review.googlesource.com/422534
Commit-Ready: Michael Tang <ntang@chromium.org>
Tested-by: Michael Tang <ntang@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@google.com>
Reviewed-by: Michael Tang <ntang@chromium.org>