Add ST to the INPUT_DEVICES list for Eve
This CL adds "st" to the list of input devices that Eve expects to
use.  This informs the build system to include the st firmware updater
tool in the image.  Without that, we can't update the firmware on any
st devices.

BUG=chrome-os-partner:61088
TEST=build_packages for Eve and with the updated touch_updater package
st_flash was built and installed.

Change-Id: I1e57cf4975b8f9c507092abb1fb3615d3d0b0630
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/422999
Commit-Ready: Tai-hsu Lin <sheckylin@google.com>
Tested-by: Tai-hsu Lin <sheckylin@google.com>
Reviewed-by: Tai-hsu Lin <sheckylin@google.com>
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>