kevin/gru: setup schedtune groups in ARC
We want to use schedtune to boost sertain threads in Android container,
let's create tuning groups there. Also expand "foreground" cpuset to
cover all CPUs.

Note that we are moving from "on-boot" to "on early-boot" because we
want tuning groups to be created before we start services.

Also, as part of this change we are extending "foreground" cpuset to run
on all CPUs.

BUG=chrome-os-partner:59485
TEST=Build and boot Kevin, verify schedtune groups are in ARC

Change-Id: Ia0dfd08f83aee56c3beabd83688404e41d33ec80
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/426758
Reviewed-by: Douglas Anderson <dianders@chromium.org>