Revert most recent TP config
The previous config + FW combo was meant to fix some noise issues, but
it appears to cause strange performance issues.  Namely latency
increases dramatically when you add a second finger to the pad.  As
such, this CL reverts back to the old version while we figure out a new
solution.  The old version of the config is susceptible to noise, but is
otherwise better than this config.

BUG=chrome-os-partner:62157
TEST=emerged and deployed on a kevin, the touchpad worked much better
after allowing the updater to apply this config.
CQ-DEPEND=CL:*318644

Change-Id: Ia314d957fbfd8be6f1f6637a5c5a89e3e799a3e9
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/430281
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>