control communications with H1 on the board level
We have been manually adding boards which are supposed to use H1 over
SPI into a case statement on cr50-update.conf file.

With a recent commit https://chromium-review.googlesource.com/c/430318
the startup file is edited at build time based on the presence of the
USE flag h1_over_spi

Let's add this USE flag to all Gru derivatives, it will be ignored on
Kevin, and is required on all others.

BUG=none
TEST=verified that withe the above mentioned patch in place invoking
     'emerge-gru chromeos-cr50' generates proper cr50-update.conf

Change-Id: I366498b2aa016cbc11219e29237b102879ab86a7
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/431051
Reviewed-by: Philip Chen <philipchen@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>