chromeos-base/modem-watchdog: add package
This CL adds a chromeos-base/modem-watchdog package, which implements a
watchdog to monitor if the built-in modem (Huawei ME936) on Kip is
successfully picked up ModemManager within a specified amount of time
after the modem appears on USB under USB configuration 3 (i.e. MBIM
mode). The watchdog will power cycle the modem upon timeout. This is a
temporary workaround of an out-of-sync issue between the modem and the
host (chromium:653979).

BUG=chromium:653979
TEST=Tested the following on a Kip with a built-in Huawei ME936 modem:
- Verified that the watchdog performs nothing and exits gracefully
  when the modem is picked up by ModemManger within the specified
  timeout.
- After reproducing the issue described in chromium:653979. verified
  that the modem is power-cycled by the watchdog and then picked up
  by ModemManger afterwards.

Change-Id: Ie5efebcc04b8aa92f8c294efc4707f3131f4e4a0
Reviewed-on: https://chromium-review.googlesource.com/431200
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>