Enable ABS_MT_TOUCH_MAJOR for Kevin's Touchscreen
This CL updates the touchscreen config for Kevin so as to enable
reporting of ABS_MT_TOUCH_MAJOR events.  Previously, these events were
omitted, and ABS_MT_PRESSURE was supplied instead, however the
performance seemed acceptable otherwise.  We asked the vendor to
just add the touch_major events since they're used to estimate
fingersize on the touchscreen.  This config seems to add it without any
noticable regressions in my testing.

BUG=chrome-os-partner:62152
TEST=emerged/deployed on my Kevin.  On reboot the config was applied
automatically and the fingersize varied when I went to
rbyers.net/paint.html and drew on the touchscreen.  The performance
seems fine otherwise, as well

Change-Id: I744b09833fc575b67166321c910699289fb8d40b
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/432576
Reviewed-by: Dennis Kempin <denniskempin@google.com>