snappy: Disable touchpad wake by default
Due to convertible wake limitations right now, we want to disable the
touchpad wake entirely until we can come up with a better way of managing
the power state of the touchpad when we transition to tablet modes
from clamshell mode when the system is asleep.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chrome-os-partner:62641,chrome-os-partner:61218
TEST=powerd_dbus_suspend with lid open in clamshell mode.
Check that touching touchpad does not wake system.
Check that transitioning to tablet mode and touching touchpad does not
wake system

Reviewed-on: https://chromium-review.googlesource.com/433328
Commit-Ready: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Signed-off-by: Harry Pan <harry.pan@intel.com>
Change-Id: I7e27afe877ee9a0c34bf1b0a1dbe78572b458c7c
Reviewed-on: https://chromium-review.googlesource.com/438026
Reviewed-by: Benson Leung <bleung@chromium.org>