reef/electro: Disable touchpad wake by default
Due to convertible wake limitations right now, we want to disable the
touchpad wake entirely until we can come up with a better way of managing
the power state of the touchpad when we transition to tablet modes
from clamshell mode when the system is asleep.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chrome-os-partner:61218
TEST=powerd_dbus_suspend with lid open in clamshell mode.
Check that touching touchpad does not wake system.
Check that transitioning to tablet mode and touching touchpad does not
wake system

Change-Id: I9b3cac87312009a37832a3cfd367d193f1b0b0eb
Reviewed-on: https://chromium-review.googlesource.com/433328
Commit-Ready: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>