Enable ABS_MT_TOUCH_MAJOR for Caroline's Touchscreen
This CL updates the touchscreen config for Caroline so as to enable
reporting of ABS_MT_TOUCH_MAJOR events.  Previously, these events were
omitted, and ABS_MT_PRESSURE was supplied instead, however the
performance seemed acceptable otherwise.  We asked the vendor to
just add the touch_major events since they're used to estimate
fingersize on the touchscreen.  This config seems to add it without any
noticable regressions in my testing.

BUG=chrome-os-partner:62152
TEST=emerged/deployed on my Caroline.  On reboot the config was applied
automatically and the fingersize varied when I went to
rbyers.net/paint.html and drew on the touchscreen.  The performance
seems fine otherwise, as well

Change-Id: Ib9b8d0f9ad7a64087361362434a6b6c0d56075ff
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/434757
Reviewed-by: Jingkui Wang <jkwang@google.com>