Update Kevin TP config to pass noise test
This CL updated the touchpad firmware config for Kevin such that it
passes our electrical noise test.  The touchpad was struggling to pass
all of our spec at once and this introduces a small regression in finger
separation (when two fingers are close together it gets hard to separate
them) but passes the noise test.  This is a tradeoff, but we've already
had users getting hit by electrical noise and that can thoroughly render
a touchpad useless, whereas finger separation can be compensated for
more easily by the user if they run into issues.  As such, this config
is preferable.  Hopefully we'll get a config that passes our whole spec
at once, but for now at least this is an improvement.

BUG=chrome-os-partner:62157
TEST=emerged and deployed on my Kevin.  The config was tested by Allion
and passed everything but the finger separation tests.  Manual testing
and feedback from dogfooders seems that this config will be preferred.

Change-Id: I38dc1755754a08f699917b11427f5201885ded06
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/434546
Reviewed-by: Jingkui Wang <jkwang@google.com>
(cherry picked from commit 18bdbc062e3674a81795087d5149a53266e22fd8)
Reviewed-on: https://chromium-review.googlesource.com/434931