Update Caroline TP config to pass noise test
This CL updated the touchpad firmware config for Caroline such that it
passes our electrical noise test.  The touchpad was struggling to pass
all of our spec at once and this introduces a small regression in finger
separation (when two fingers are close together it gets hard to separate
them) but passes the noise test.  This is a tradeoff, but we've already
had users getting hit by electrical noise and that can thoroughly render
a touchpad useless, whereas finger separation can be compensated for
more easily by the user if they run into issues.  As such, this config
is preferable.  Hopefully we'll get a config that passes our whole spec
at once, but for now at least this is an improvement.

BUG=chrome-os-partner:62157
TEST=emerged and deployed on my Caroline.  The config was tested by Allion
and passed everything but the finger separation tests.  Manual testing
and feedback from dogfooders seems that this config will be preferred.

Change-Id: Ib25b6b7effef568c8135be94f423935b4663fc5e
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/434879
Reviewed-by: Jingkui Wang <jkwang@google.com>
(cherry picked from commit 0f2a464886c8437a2a394c27b64d813bc8f13b1c)
Reviewed-on: https://chromium-review.googlesource.com/435844