lakitu: Expand bindmount@.service into .mount units
bindmount@.service is a template unit to mimic what a .mount unit does.
Unfortunately the RequiresMountsFor directive doesn't take Specifiers as
others may do, in the version of systemd we use.

This change tries to expand it into individual .mount units instead.

The advantage is that the "Default Dependencies" for .mount units should
just work - we don't have to craft them by hand. For example systemd
automatically adds a Require and After dependency of .mount units for
all parent paths so we don't have to do RequiresMountsFor.

The downside is that we now need two units to do a bind mount - we used
to, too. :)

BUG=b:33245793
TEST=trybot

Change-Id: Iaa891319751464220bec8e3b3bcfe26ec7d20668
Reviewed-on: https://chromium-review.googlesource.com/435468
Commit-Ready: Daniel Wang <wonderfly@google.com>
Tested-by: Daniel Wang <wonderfly@google.com>
Reviewed-by: Andrey Ulanov <andreyu@google.com>