audio-config: Change hacks by customization_id to "mosys platform chassis".
Device difference should be identified by the new unified "chassis id"
provided by mosys. In the future the command may return values from
hardware GPIO settings, but for legacy devices for backward
compatibility the command is returning SERIES part of VPD
customization_id.

When customization_id is not provided, "mosys platform chassis" will
return empty string. As a result, for some boards which expect to have a
"default" value (i.e., RELM or REKS), we have to put that in the
scripts.

Also, RELM was using customization_id in a wrong approach - that it
tries to find SERIES from LOEMID part. Ideally this should be fixed in
RELM's "mosys chassis" command, but during the migration we want to do
the remapping first in cras-config scripts.

BUG=chromium:687149
TEST=manual:
     vpd -d customization_id; dump_vpd_log --force; reboot;
     get_device_config_dir # prints /etc/cras/
     vpd -s customization_id=A-B; dump_vpd_log --force; reboot;
     get_device_config_dir # prints /etc/cras/B

     (reks)
     vpd -d customization_id; dump_vpd_log --force; reboot;
     get_device_config_dir # prints /etc/cras/REKS1
     vpd -s customization_id=A-B; dump_vpd_log --force; reboot;
     get_device_config_dir # prints /etc/cras/B

     (relm)
     vpd -d customization_id; dump_vpd_log --force; reboot;
     get_device_config_dir # prints /etc/cras/RELM
     vpd -s customization_id=A-B; dump_vpd_log --force; reboot;
     get_device_config_dir # prints /etc/cras/B
     vpd -s customization_id=ACERTIFA; dump_vpd_log --force; reboot;
     get_device_config_dir # prints /etc/cras/ACERTIFA
     vpd -s customization_id=ACERTIFA; dump_vpd_log --force; reboot;
     get_device_config_dir # prints /etc/cras/DASANI

Change-Id: I846dbdaa9bd02c130f31f5d251bf2dc7a3a46aac
Reviewed-on: https://chromium-review.googlesource.com/438025
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>