Revert Caroline TP config (7B1A94 to BAA5E9)
The previous firmware config for the touchpad (checksum 7B1A94) was a
compromise.  It passed the electrical noise test, but had a few known
minor issues surrounding finger separation.  That tradeoff seemed worth
it, but we've had reports coming in from the field that claim the
touchpad is reporting hovering fingers as if they're touching the
touchpad.  I've reproduced this myself, and you can indeed get it to
report fingers 1-2mm above the pad as long as they've already touched
the surface.  This is causing issues with scrolling and flinging and
needs to be reverted.

BUG=chrome-os-partner:62657
TEST=emerged and deployed on my Caroline.  The touchpad got updated
successfully on the next reboot and I was unable to reproduce the
hovering effect with config BAA5E9.

Change-Id: Id00c3a4e11d89235d6f09a8ae4473fb7ead4ed1f
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/438645