Revert Kevin TP config (93E2A4 to BAA5E9)
The previous firmware config for the touchpad (checksum 93E2A4) was a
compromise.  It passed the electrical noise test, but had a few known
minor issues surrounding finger separation.  That tradeoff seemed worth
it, but we've had reports coming in from the field that claim the
touchpad is reporting hovering fingers as if they're touching the
touchpad.  I've reproduced this myself, and you can indeed get it to
report fingers 1-2mm above the pad as long as they've already touched
the surface.  This is causing issues with scrolling and flinging and
needs to be reverted.

BUG=chrome-os-partner:62657
TEST=emerged and deployed on my Kevin.  The touchpad got updated
successfully on the next reboot and I was unable to reproduce the
hovering effect with config BAA5E9.

Change-Id: I32091f0c84e3c8dac07d0307c6bfc8c4cf9ee09c
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/438527
Reviewed-by: Jingkui Wang <jkwang@google.com>
Reviewed-on: https://chromium-review.googlesource.com/456137
Tested-by: Jongpil Jung <jongpil19.jung@samsung.com>
Commit-Queue: Jongpil Jung <jongpil19.jung@samsung.com>