moblab: drop jabra-vold from the build
This daemon is used for CfM to work with a jabra speakerphone.
It also requires some custom JS running on the system.  Neither
of these apply to moblab.  I suspect this is just a copy & paste
error from the initial port, so drop the dep.

BUG=None
TEST=precq is happy

Change-Id: I93f299c08e2d00d6f6fcd348ef848eeebc9d366a
Reviewed-on: https://chromium-review.googlesource.com/448258
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Michael Tang <ntang@chromium.org>