[moblab] Rotate the apache logs to limit the disk space used.
Partners running storage qual are filling up internal disk, one
major culprit is the apache access log, configure apache to
rotate and limit the size of the log.

Only noted side effect is the logs are initially owned by root not
apache, however after a reboot they become apache, I suspect this
behavior is as before but just noted because it was not what I expected.

TEST=manual tests on a moblab
BUG=chromium:701467

Change-Id: I2395a899f1dcf23932a4b2c18b4521902eb23831
Reviewed-on: https://chromium-review.googlesource.com/455517
Reviewed-by: Michael Tang <ntang@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>