Reduce Kevin/Caroline touchpad pressure hysteresis
This CL updates the touchpad firmware config for Kevin/Caroline to have a
smaller "pressure hysteresis" which is the amount of pressure change that
must be detected before reporting it.  Previously it was a bit too high
so this config is itentical apart from reducing this value down to 1.
This means pressure changes are more readily reported by the touchpad
and improves our gesture library's performance.  Hysteresis is usally
added to reduce jitter, but out gestures library can handle jitter on
its own.

BUG=b:36240752
TEST=emerged/deployed on a Kevin and a Caroline.  The updater sucessfully
applied this new confi.  evtest confirms that the pressure value changes
in a much more fine-grained steps than before and basic usage/performance
appears the same as before which is consistent with the config being
otherwise unchanged.

Change-Id: Ie40e83add1fec8c5c9a533f79398edba90a3eb39
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/458260
Reviewed-by: Marco Chen <marcochen@chromium.org>