lakitu: make cloud-init wait for 'user-data' to be available
Because of delay in metadata endpoints becoming active, cloud-init
often ends up running when 'user-data' is not accessible.
This patch adds an ExecStartPre= condition to cloud-init.service
that waits for some time (60 mins) for user-data key to be available.
This is a workaround until a better signal for metadata reachability
is available.

RELEASE_NOTE=Ensure cloud-init waits for user-data to become accessible.

CQ-DEPEND=CL:*364093
CQ-DEPEND=CL:497330
BUG=b:34390380
BUG=b:37570951
TEST=verified that cloud-init blocks until user-data is available and
proceeds successfully once it is.

Change-Id: I62b037b90d0f337b8cb13d654ccfa8c5a0bc5ac4
Reviewed-on: https://chromium-review.googlesource.com/487867
Commit-Ready: Aditya Kali <adityakali@google.com>
Tested-by: Aditya Kali <adityakali@google.com>
Reviewed-by: Aditya Kali <adityakali@google.com>
Reviewed-by: Amey Deshpande <ameyd@google.com>