[moblab] Set an Apache define to reflect the apache is on moblab.
In other configuraitons we need to be able to change what the main
lab setting is.  For cases where apache configuraition does not
allow settings to be overridden ( line WSGI ) then use this define
to add configuration logic so the correct settings are used for
the correct platform.

TEST=Manual testing on moblab
BUG=chromium:716181

Change-Id: I542fd5df1119df413db3613404239de3d445284f
Reviewed-on: https://chromium-review.googlesource.com/499087
Commit-Ready: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Michael Tang <ntang@chromium.org>