lakitu: set CONFIG_IKCONFIG to 'y'
This is useful for building kernel modules in containers. Before this
patch, users would need to `modprobe configs` for /proc/config.gz to
appear on the system.  With this patch, /proc/config.gz exists upon
boot.  The disk overhead in vmlinuz size is 0.4% (4470320 bytes before
this patch, and 4488752 with this patch).

BUG=b:38045909
TEST=Booted an instance and verified presence of /proc/config.gz.

Change-Id: I3cbc01a175bf8ec83cc83d7c0a297112eb55ea4e
Reviewed-on: https://chromium-review.googlesource.com/505456
Commit-Ready: Amey Deshpande <ameyd@google.com>
Tested-by: Amey Deshpande <ameyd@google.com>
Reviewed-by: Aditya Kali <adityakali@google.com>
Reviewed-by: Andrey Ulanov <andreyu@google.com>