baseboard-gru: Turn on the serial console again
The serial driver is always enabled in the kernel on gru devices.
That means we have the ability to always enable login via the serial
port.

Apparently the serial console isn't intended to be left on for
non-debug mode.  Now that we've fixed our init script to avoid
spawning the agetty in debug mode, there should be no problems turning
this on always for gru.

CQ-DEPEND=CL:506770
BUG=chromium:722913
TEST=build_image

Change-Id: I21816f620c1125f9ce7c11944f9444e4807e0eb3
Reviewed-on: https://chromium-review.googlesource.com/506807
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Nick Sanders <nsanders@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>