[moblab] Set CROS_CACHEDIR for apache service.
The apache user can not currently share the same gsutil installation
as moblab user since the gs code attempts to update the installation
and the directory is owned by one or the other users.

Ideally the gs code would be updated to handle this case well, however
for now make a separate gsutil install just for apache, it is only 20M
of disk space.

BUG=chromium:723870
TEST=manually on moblab
CQ-DEPEND=CL:510014

Change-Id: I1c29a5471136694c628dbe15ddce6782ac25773d
Reviewed-on: https://chromium-review.googlesource.com/513395
Commit-Ready: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Jinsong Mu <jinsong@google.com>
Reviewed-by: Keith Haddow <haddowk@chromium.org>