Update Caroline/Kevin's TS config to detect noise
This config is a slight modification of the previous one given to us by
the touchscreen vendor in response to some units having defective
connectors.  Essentially on some bad units the connector is not totally
reliable and sometimes it loses connection briefly which can cause the
touch controller to get bad "baseline" or "reference" readings.  When
it gets a bad baseline all bets are off performance-wise until it
resets.

This config turns on a feature in the touch controller that tries to
detect this condition by looking for any large negative values (<-960 in
this case) in the "deltas" and triggers a full reset when one is
detected.  These values should always be either small negative numbers
near zero or large positive numbers.  Atmel tells use that the only
reason for a large negative value is a bad baseline.  So this should
mitigate some of the effects from a bad connector by resetting the touch
controller in this situation.

Users may notice a ~1s interruption in touch during the reset, but this
is preferable to the touchscreen remaining in a bad state until reboot.

BUG=b:35775059
TEST=built and deployed onto my Kevin and a Caroline.  Performance
appears unchanged, but I was not able to reproduce the issue (it's
*very* infrequent).  Atmel assures me that this has been heavily
tested in their labs and passed all their testing and that there is
very little risk here.  I've seem no issues myself, so I'm inclined to agree.

Change-Id: Ide7e2d5b2f5774b3ae9dffe0b7be1947af6840e3
Signed-off-by: Charlie Mooney <charliemooney@google.com>
Reviewed-on: https://chromium-review.googlesource.com/545080
Commit-Ready: Charlie Mooney <charliemooney@chromium.org>
Tested-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-by: Jingkui Wang <jkwang@google.com>