Delete pyro/sentry specific activate_date scripts
The only delta with the pyro/sentry impl is that is attempted to set the First
Use Date (FUD) on the battery; however, the fuddata was empty in the
script anyway, which caused ectool to error out, but the script
passively ignored the error anyway.

So the net of this customization was nothing.  Additionally though, this
script has multiple other bugs that were fixed in the base impl of the
script.

This breaks the delete into 2 steps so that there are no file conflicts with the
incremental paladin builder.  It first puts hardblockers inplace to the old virtual
packages so that they are unmerged and then adds a new dep to the base impl.

On the 2nd step, these virtual impls will be deleted altogether.

BUG=chromium:748153
TEST=locally on sentry and using the new functional test
platform_ActivateDate

Change-Id: Ia491126753a04c886d353b5f16b61aefb78661ee
Reviewed-on: https://chromium-review.googlesource.com/587400
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>