[moblab] Add an include to a dhcp reservation file for testing.
In the lab we see devices not getting the same IP address after
reboot.

Change the dhcp config to include a file that can optionally
contain dhcp reservations by mac address.

The reservations will be generated manually in the lab for testing.

If all goes well a UI option to add dhcp resevations when you are
managing a DUT may be added later and the file moved so it will
not survive a powerwash.

BUG=chromium:748846
TEST=local build on moblab and will complete a trybot job

Change-Id: I1144412d4dbc1f8fc1939530b798c3c66bee5b58
Reviewed-on: https://chromium-review.googlesource.com/588068
Commit-Ready: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Michael Tang <ntang@chromium.org>