chromeos-overlay: Remove need-arc-migration-policy-check USE flags.
This reverts commit 3aceabc2628edabf2e8c0442137060a327e1d108.

Reason for revert: These overlays are using device policy in M-60 to
decide if ecryptfs to ext4
migration option is given to the user. Starting from M-61 that
functionality is covered by an
user policy. To avoid possible conflicts, the functionality that uses
the device policy has to
be rolled back starting from M-61.

BUG=chromium:725493
TEST=None
Change-Id: Id88dca04ebe8a1de41c6dd215022e94ace7b8151
Reviewed-on: https://chromium-review.googlesource.com/596287
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>