moblab: Allow explicitly setting network interface for private network
moblab uses one of its two (or more) network interfaces for a private
network to be used with its DUTs. So far, a heuristic was used to
determine which network interface to use for this purpose. This CL adds
a way for a moblab device to explicitly state the mac address of the
interface to be used for this purpose.

BUG=chromium:725620
TEST=Test network setup in the moblab VM.
     + sanity moblab_RunSuite passes in the lab (not using the new
     feature)

Change-Id: I17f1bbdfdaf4b53fb4bd791a8262113380753460
Reviewed-on: https://chromium-review.googlesource.com/604927
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>