cave: audio-config: Decrease lower end speaker volume
There have been many people complaining that cave doesn't have volume
settings low enough for listening in a quiet environment. I noticed
this as well because cave is my regular device.

Set the volume levels for the last part of the speaker volume curve so
they drop off much faster and volume four, the lowest reachable from the
buttons, is low enough when in a quiet room.

Change-Id: I9ca8537b7daedf99dd8b5e62bf79fd3fac8963ec
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/696424
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Yomyung Leem <yungleem@chromium.org>