Soraka: HiFi.conf: Support front and rear mics.
There are two mics on the soraka device; one is for user facing and the
other is for wide facing. Hence the SectionDevice of Front Mic and Rear Mic
are added in the HiFi.conf for cras to support them.

On the other hand, correct the device number of Capturing device of mics.

BUG=b:65148117
TEST=Manually test on the DUT.

Change-Id: I3f273fad1122678d736227995ba204501038efc9
Reviewed-on: https://chromium-review.googlesource.com/727789
Commit-Ready: Marco Chen <marcochen@chromium.org>
Tested-by: Marco Chen <marcochen@chromium.org>
Reviewed-by: Marco Chen <marcochen@chromium.org>