lakitu: make systemd use the legacy cgroup hierarchy by default
On newer kernel (ex: 4.12) systemd tries to use cgroup-v2 hierarchy for
the systemd controller (this is the "hybrid" mode for systemd).
It is unclear if this is useful for us right now or if this will lead to any issues.
For now, to avoid unintended regression, force systemd to use the legacy
cgroup hierarchy for everything which is the default behavior on
current 4.4 kernel. This is achieved by setting:
  systemd.unified_cgroup_hierarchy=false
  systemd.legacy_systemd_cgroup_controller=true

on the kernel command-line.
See https://www.freedesktop.org/software/systemd/man/systemd.html for
details.

This change is a no-op with the current 4.4 kernel.

BUG=b:68010453
TEST=cloud_ProcessTree test fails because of this mismatch on newer
kernel. With this patch, it passes. Also manually verified that legacy
hierarchy is used.

  # Without this patch on 4.12 kernel (note the "cgroup2" fstype):
  $ cat /proc/mounts | grep "cgroup/systemd"
  tmpfs /sys/fs/cgroup tmpfs ro,nosuid,nodev,noexec,mode=755 0 0 cgroup /sys/fs/cgroup/systemd cgroup2 rw,nosuid,nodev,noexec,relatime 0 0

  # With this patch on 4.12 kernel:
  $ cat /proc/mounts | grep "cgroup/systemd"
  cgroup /sys/fs/cgroup/systemd cgroup rw,nosuid,nodev,noexec,relatime,xattr,release_agent=/usr/lib/systemd/systemd-cgroups-agent,name=systemd 0 0

RELEASE_NOTE=None

Change-Id: I00e45dd7fa3d72bb0499bed421e3e003804945c1
Reviewed-on: https://chromium-review.googlesource.com/735979
Commit-Ready: Aditya Kali <adityakali@google.com>
Tested-by: Aditya Kali <adityakali@google.com>
Reviewed-by: Edward Jee <edjee@google.com>
Reviewed-by: Daniel Wang <wonderfly@google.com>